requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
babel==2.16.0
beautifulsoup4==4.12.3
certifi==2024.8.30
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-redis==5.4.0
django-siteajax==1.0.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
phonenumbers==8.13.48
pillow==11.0.0
postgres==4.0
psycopg2-binary==2.9.10
psycopg2-pool==1.2
python-dotenv==1.0.1
python-gettext==5.0
python-telegram-bot==21.7
pytz==2024.2
redis==5.1.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
whitenoise==6.8.2